+++ Assets/LDtkUnity/CHANGELOG.md
@@ -1,6 +1,9 @@
 # 4.1.1
-###### January 1, 2023
+###### January 1, 2024
 - Updated json support & samples for [LDtk 1.5](https://ldtk.io/json/#changes;1.5.0)
+- Added support for the additional Table-of-content data from LDtk 1.5
+  - Except the `Fields`; will come later 
+- Deprecated `LDtkTableOfContents.GetEntities`, use `GetEntry` instead
 
 # 4.1.0
 ###### October 23, 2023